[namespacing] update link-to-external-component to be called link-to-external so references to it are not confused
