+++ addon/-private/engine-ext.js
@@ -1,6 +1,6 @@
 import Application from '@ember/application';
 import EngineScopedLinkComponent from '../components/link-to-component';
-import ExternalLinkComponent from '../components/link-to-external-component';
+import ExternalLinkComponent from '../components/link-to-external';
 import Engine from '@ember/engine';
 
 Engine.reopen({
